services/image-service/src/domain/entities/ChatMessage.py
from datetime import datetime
from typing import Optional, Literal
from pydantic import BaseModel, Field
from bson import ObjectId


class ChatMessage(BaseModel):
    id: Optional[str] = Field(default=None, alias="_id")
    image_id: str
    user_id: str
    role: Literal["user", "assistant"]
    content: str
    timestamp: datetime = Field(default_factory=datetime.utcnow)

    model_config = {
        "populate_by_name": True,
        "arbitrary_types_allowed": True,
        "json_encoders": {ObjectId: str},
    }


