services/image-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
aiofiles==23.2.1
python-magic==0.4.27
requests>=2.31.0 
openai>=1.43.0
boto3>=1.34.0
botocore>=1.34.0